Update Euclidean_Algorithm.md
Please make it look like this. and this pull request dont merge it. It is just for the changes and a material which may act as a guide. Include that improvement etc part in extended one. Tell the advantage and disadvantage and that stuff. Try to write relevant things in point as it can help in a better way.
Thanks for contributing :smiley: